leap: first pass at cleaning ncsgl_defs.h of any int64/uns64 for stdint.h (#655)
  * first validation pass, swap ncsgl_defs.h to C99 types
  * Adapt printfs to stdc format macros when needed
  * global CXXFLAGS to support -D__STDC_FORMAT_MACROS
  * remove NCS_UNS64_DEFINED and NCS_64BIT_DATA_TYPE_SUPPORT
  * remove dead code

+++ osaf/libs/core/include/ncsgl_defs.h
@@ -43,6 +43,8 @@
 #ifndef NCSGL_DEFS_H
 #define NCSGL_DEFS_H
 
+#include <inttypes.h>
+
 #ifdef  __cplusplus
 extern "C" {
 #endif
@@ -65,12 +67,12 @@ extern "C" {
 	typedef unsigned char uns8;	/*  8-bit */
 	typedef unsigned short uns16;	/* 16-bit */
 	typedef unsigned int uns32;	/* 32-bit */
-	typedef unsigned long long uns64;	/* 64-bit */
+	typedef uint64_t uns64;	/* 64-bit */
 
 	typedef signed char int8;
 	typedef signed short int16;
 	typedef signed int int32;
-	typedef long long int64;
+	typedef int64_t int64;
 
 	typedef float ncsfloat32;
